Bypass prompts on gopass insert --force
This makes it easier to write an automated import script from other
password managers. In particular, I tried replacing pass with gopass in

https://git.zx2c4.com/password-store/tree/contrib/importers/lastpass2pass.rb

and it failed because it got an EOF at the Y/n query. With this change
and an added -f option in the script I could import a lastpass file.