Handle borked return value for terminal.Restore
Until recently (go 1.8 still seems to be doing it the old way)
`terminal.Restore` returned an error that was distinctly non-go-like.

This resulted in this error message whenever the terminal was set to
no-echo and then restored:

> Failed to restore terminal: errno 0

It's actually returning the `syscall.Errno` returned by
`syscall.Syscall6(..)`.  When it's 0 everything was successful,
otherwise there's a problem.

`terminal.Restore` is returning it as a value of type `error`, which
makes it hard to test for zero-ness.

This odd behavior has been [fixed in this commit][fix].

I still see this problem when I build gopass with go@1.8, so apparently
the [fix] did not make the cutoff.

[fix]: https://github.com/golang/crypto/commit/c3b1d0d6d8690eaebe3064711b026770cc37efa3