Adding a setting to avoid showing password
By default, especially when usernames or comments in password entry are
useful, it may be dangerous to display the password on the terminal.
This commit add a setting to handle this, as well as a way to force
current behaviour.
The default behaviour is changed to this more sane and secure way.
Unit tests for the feature are passing.