Correcting the dmenu cmd in the docs (#441)
* Correcting dmenu call

Since if the setting `safecontent` is set to `true`, it won't work like this. In order to be sure to get the password on the first line no matter what, you need to use the `-f` flag. (Which do nothing if `safecontent` is set to `false`)

* Adding some details

So it's all explained now.