Fixes #667 and correct a few details
This primarily adds the config option requested in #667.
I've added back the notion or erroring in case of a show action on an
entry with only a password and the safe content option enabled
Also I've fixed the way the String() function of the StoreConfig forgot
to display the Notification value in its output.