Unbreak Makefile (#707)
* Makefile: fix pointless repetition of a custom command

Drop unneeded variables, reorganize the precedence of .PHONY in
accordance with standard style (this has no reason to be one of the
first things people see).

* Makefile: fix completely broken target ordering.

Previously, all targets were .PHONY targets, which meant that on every
invocation of `make`, all targets were unconditionally rebuilt. For
example, running `make && make install` would build gopass, then rebuild
it in order to build the completions.

The solution is to properly use file-based dependencies. The gopass
target is used to generate the gopass binary, and properly depends on
all its source files. The completion files are generated using a
Makefile pattern, and regenerated whenever the gopass binary is. Various
other commands are modified to depend on the binary instead of the build
target (which is now an alias to build the gopass target).

While I am at it, the default `make` target (that is to say, the first
listed rule) should *not* be the one that runs all of the project's own
internal unit testing and coverage commands.
So, make the default `make all` target build the project and generate
completion files, and use the `make travis` target for, well, making
travis-related things.