Follow the rsync convention for the source directory in cp/mv commands (#1055)
When moving/copying folders, gopass currently enforces a trailing slash
on the source folder. This makes it impossible to rename a folder.
Rsync re-creates the source folder at the destination, if the source
folder has _no_ trailing slash. Otherwise the folder contents are
flattened at the destination location. With that convention it will be
possible to also rename folders.

Signed-off-by: Cornelius Weig <cornelius.weig@tngtech.com>