completion: fix zsh completion for folders
For `insert`, `generate` and `list`, the folder completion doesn't
work as expected. First, we should continue completion even if we
already have a `/` in the current name. Otherwise, we can only
complete the first level. Then, don't specify `/` as a separator for
`_values`, this is for specifying multiple values.

Moreover, allow completing for passwords for `generate` (ability to
change a password) and for `insert` (ability to append data).

Signed-off-by: Vincent Bernat <vincent@bernat.ch>