use double quotes for git-credential helper
We use the "!" command in the gitconfig to start a subshell and execute
a process, therefore we need to double quote the command in the
git configuration file. Otherwise the file will not get interpreted as
command and will be used directly as a helper. This commit makes sure,
that we use double quotes in our git configuration file and our gopass
help prompt.

Signed-off-by: Christian Rebischke <chris@nullday.de>