Let find support the -force flag (#1251)
Currently when the safecontent option is enable and the find action is
used and returns a single exact match, it will say: "Found exact match
in 'path'. No safe content to display, you can force display with show -f."

This is not accurate, since the -f flag is not supported by the find
action. This commit adds support for it so that it works as intended.

Adding some tests as well.

Signed-off-by: Yolan Romailler <yolan@romailler.ch>