Replace gopass agent by askpass implementation (#1294)
* Refactor agent - move out cache

This commit moves the cache implementation in the agent into it's own
package.

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>

* Replace gopass agent by askpass implementation

This commit removes the incomplete gopass agent with a simple in-process
ask password implementation.

The agent did never really work well and isn't strictly needed for
most use cases anyway. In process bulk operations are still supported
with the new implementation. Only use cases where an external
process calls gopass repeadetly are not well supported by this.

If we decide to bring back agent support, we should consider using
the extensible SSH agent protocol for this, simliar to what
github.com/42wim/age and github.com/42wim/sagent are doing.

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>