Add age support (#1337)
This is a proof of concept for age-encryption.org support.

THIS IS NOT SAFE TO USE, YET.

It adds support for an age encrypted backend that supports both native
age as well as ssh keys for encryption. It also supports encrypting to
GitHub usernames, if the publish their ssh public keys.

The native key pair is automatically generated on first use, but not yet
encrypted. So handle with care.

Unfortunately age does not expose an API (yet), so this relies on the
age and age-keygen tools being in the users PATH.

By design age will not support some of the gopass features, e.g. listing
recipients of a secret or handling key metadata (there is none).

This is only intended as a proof of concept, to experiment with age
and provide a base to build future commits upon.

If this experiement is successfull it will likely replace the xc
backend.

Fixes #1197

RELEASE_NOTES=[ENHANCEMENT] Add preliminary age encryption support.

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>