Encrypt age keypair
This works, but has awful UX. Due to how gopass works internally
most operations require the keypair password to be entered multiple
times.

This should offer a similar level of protection as the other
backends but at a very high cost.

Unfortunately as of today age can neither be used as a library
nor can the password be passed - it can only read directly from
the terminal.

Since we expect age to offer a Go API soon, I'll add this as
an intermediate step. Once the API is available we should
be able to cache the password so it only has to be entered
once per gopass invocation (much like we do with xc today).

Fixes #1344

RELEASE_NOTES=n/a

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>