Use the new Age API
This migrates from calling age as an external binary to using
age as a library using it's new API.

The implementation is still ugly, hacky and likely buggy.
But it should be as secure as age and provides a nice proof
of concept. It also has an encrypted keyring and when using
the (highly experimental) ondisk backend it's possible to
have an fully encrypted store without any external dependencies.

RELEASE_NOTES=n/a

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>