Fix panic in config and cleanup
Fixes #1458
Fixes #1459

This is fixing a panic in config that was not catched by the tests, it
augments the tests to catch it and it cleans up the config code that
still had some code from the time where we supported per-mount
configuration. Documentation is updated as well to reflect the change
made in #1391 that is dropping some config options.

It is also adding a test to detect regressions in the ways mount
shadowing works.

RELEASE_NOTES=[BUGFIX] Fix config panic with mounts

Signed-off-by: Yolan Romailler <yolan@romailler.ch>