fixed MacTootls GPGSuite path problem in jsonapi-wrapper
macos-users not using homebrews gpg, but the MacTools GPGSuite won't be
able to use the gopassbridge-browser-plugin as the gopass-jsonapi cannot
locate the gpg-binary.  this fix adds the MacTools GPGSuite bin folder to
the PATH. more specific description of the issue can be found under
https://github.com/gopasspw/gopassbridge/issues/151

RELEASE_NOTES=[BUGFIX] fixes gopass-jsonapi for MacTools GPGSuite users.

Signed-off-by: Stefan Geisbacher <stefan.geisbacher@gmail.com>