Fixing Windows tests
RELEASE_NOTES=[ENHANCEMENT] Full windows support

Windows tests are no longer allowed to fail now since they all pass.
Notice the exception for jsonapi which is considered a malware for some
reason and is therefore excluded on windows.

We will be able to remove the exception once jsonapi will be moved to
its own repo.

Signed-off-by: Yolan Romailler <yolan@romailler.ch>