Don't prompt to retype password unnecessarily with age backend (#1983)
When gopass can't find a pinentry binary to use for prompting for a
password to unlock an age keyring, it uses its own barebones fallback
pinentry mechanism to prompt for a password in the terminal.  This
fallback pinentry always asked the user to retype their password and
required the password match, even when decrypting a file using an
already-existing password.  I've updated it to only prompt for a repeat
when necessary, and also made the password prompte messages less
awkward.

RELEASE_NOTES=[BUGFIX] Don't prompt to retype password unnecessarily
with age backend when pinentry binary is unavailable.

Signed-off-by: Faye Duxovni <duxovni@duxovni.org>