Remove recipients based on Key ID
This PR removes some unnecessary logic in the recipients removal and
adds a shortcut to remove recipients based on the Key ID instead of
looking up key fingerprints over and over again. Also it fixes an
issue where subkeys of the same key wouldn't be displayed in gopass
recipients even if they were actually present in the .gpg-id files.

Fixes #2416

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>