[feat] Allow supression of password generation in create templates (#2821)
* [feat] Allow supression of password generation in create templates

This PR adds a new always_prompt boolean that allows suppressing the
question that ask for generating a password in create templates. Use
this is you always want to force entering the credentials, e.g. if
you use a specific template for vendor-supplied credentials.

Fixes #2819

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>

* Remove failing integration test

The messages have changed and the removed one is not expected anymore.

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>

* Fix lint check

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>

---------

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>