[feat] Add an multi-line input type to the create wizard
This input type will spawn an editor, evaluate templates
and set that as the body of the created secret.

While at it the commit also tries to make the wizard more
robust in case templates can not be written. Then it will
fall back to the built-in default templates.

Fixes #2795

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>