[bugfix] Default to true for core.exportkeys even in substores
This PR changes the default for core.exportkeys from false to true
in mounted substores to match the default of the global root store.

It also refactors and simplifies the config package a little bit
by removing special typed lookup methods and replacing them with
conversion helpers that can be applied to any string.

Fixes #2830

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>