feat: add a new build job: container
This change adds a new job in `//.github/workflows:build.yml` named
`container`, to build the container image. This is being added as a
quick hack to fix a critical issue with this repository (in regards to
its container image), namely, that it is not built except for releases.

This means that both source code and configuration changes that break
the container build will not be caught until an attempt to create a new
release is made (more specifically, given the current configuration of
the `//.github/workflows:container.yml` pipeline, whenever a new ref is
pushed up matching `refs/tags/v*`).

An example of this is a recent change to the required version of `go`,
made in commit 4c2caf3e9b6f1c65857c0d33208e3273c2ab7eb3 (ca. October
2024), which did not udpate the base `golang` image referenced in
`//:Dockerfile`.

As a requirement to pass the new pipeline, this change also updates the
`golang` image we source in `//:Dockerfile`, in order to fix the build.

Signed-off-by: sudoforge <no-reply@sudoforge.com>