chore: remove unecessary variable: REGISTRY
This change removes the `REGISTRY` variable in the `container` job
defined in `//:.github/workflows:build.yml`. This doesn't matter on a
technical level because the image isn't pushed anywhere, but uh... I
caught the typo and figured getting rid of REGISTRY would be a good
way to make it clear that the image isn't pushed (and get rid of my
typo).

Signed-off-by: sudoforge <no-reply@sudoforge.com>