[fix] Include git commit hash in tarballs (#3124)
* [fix] Include git commit hash in tarballs

This change should allow builds made from a release tarball to
accurately display the git commit hash it was built from.

It does so by rendering the current short hash when packaging
the tarball. This workaround is necessary since the tarball
won't include the .git directory so we can't extract the
commit information from it.

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>

* [fix] Fix TestVersionPrinter

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>

* [fix] Fix spacing

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>

---------

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>