[feat] Add Jujutsu storage backend (#3202)
This commit adds a new experimental storage backend for the Jujutsu
VCS. It currently is a wrapper around Git and does not add too much
value by itself, but provides an interface that is supposed to be
more stable than git and might eventually evolve to support a
custom storage backend as well. Also I want to test it.

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>