[chore] Add capability and vulnerability checks (#3266)
This commit adds capslock and govulncheck to help us identify
vulnerabilities and possible supply chain attacks through added
capacilities in dependencies. The signal for this repop is low since
we have to rely on a large number of direct filesystem access
and subprocess executions but it seems worth a try.

Fixes #3254

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>